Switch to new vector module
I don't think I made a note about it, but one of the last couple of commits
fixed the underlying root issue of why we haven't upgraded before. For some
reason we were caching blank tile holder surfaces, and not being careful about
keeping results from that function duplicated. Fixed by removing caching in
the first place. Net effect prior to fixing was that touching that code was
a dice roll of hitting cache or not, leading to all factory tiles being displayed
as blank once one of them was updated. This ment that previous attempts to
replace vector module with new version lead to baffling cases and much confusion,
and so previously I decided to keep the old version around until I could find
out what was going on. Now that underlying issue is fixed, we can now get
rid of old verison of vector module!